Refactor UiaCacheRequest into hand-written default constructor
