+++ src/UIAutomation/UiaOperationAbstraction/UiaTypeAbstractionEnums.g.h
@@ -1,0 +1,142 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// WARNING: Please don't edit this file. It was generated.
+
+#pragma once
+
+    using UiaActiveEnd = UiaEnum<
+        ActiveEnd,
+        winrt::Microsoft::UI::UIAutomation::AutomationActiveEnd,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteActiveEnd>;
+    using UiaAnimationStyle = UiaEnum<
+        AnimationStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationAnimationStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteAnimationStyle>;
+    using UiaAnnotationType = UiaEnum<
+        int,
+        winrt::Microsoft::UI::UIAutomation::AutomationAnnotationType,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteAnnotationType>;
+    using UiaBulletStyle = UiaEnum<
+        BulletStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationBulletStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteBulletStyle>;
+    using UiaCapStyle = UiaEnum<
+        CapStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationCapStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteCapStyle>;
+    using UiaCaretBidiMode = UiaEnum<
+        CaretBidiMode,
+        winrt::Microsoft::UI::UIAutomation::AutomationCaretBidiMode,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteCaretBidiMode>;
+    using UiaCaretPosition = UiaEnum<
+        CaretPosition,
+        winrt::Microsoft::UI::UIAutomation::AutomationCaretPosition,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteCaretPosition>;
+    using UiaControlType = UiaEnum<
+        CONTROLTYPEID,
+        winrt::Microsoft::UI::UIAutomation::AutomationControlType,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteControlType>;
+    using UiaDockPosition = UiaEnum<
+        DockPosition,
+        winrt::Microsoft::UI::UIAutomation::AutomationDockPosition,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteDockPosition>;
+    using UiaExpandCollapseState = UiaEnum<
+        ExpandCollapseState,
+        winrt::Microsoft::UI::UIAutomation::AutomationExpandCollapseState,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteExpandCollapseState>;
+    using UiaFlowDirections = UiaEnum<
+        FlowDirections,
+        winrt::Microsoft::UI::UIAutomation::AutomationFlowDirections,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteFlowDirections>;
+    using UiaHeadingLevel = UiaEnum<
+        HEADINGLEVELID,
+        winrt::Microsoft::UI::UIAutomation::AutomationHeadingLevel,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteHeadingLevel>;
+    using UiaHorizontalTextAlignment = UiaEnum<
+        HorizontalTextAlignment,
+        winrt::Microsoft::UI::UIAutomation::AutomationHorizontalTextAlignment,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteHorizontalTextAlignment>;
+    using UiaLandmarkType = UiaEnum<
+        LANDMARKTYPEID,
+        winrt::Microsoft::UI::UIAutomation::AutomationLandmarkType,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteLandmarkType>;
+    using UiaLiveSetting = UiaEnum<
+        LiveSetting,
+        winrt::Microsoft::UI::UIAutomation::AutomationLiveSetting,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteLiveSetting>;
+    using UiaNavigateDirection = UiaEnum<
+        NavigateDirection,
+        winrt::Microsoft::UI::UIAutomation::AutomationNavigateDirection,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteNavigateDirection>;
+    using UiaOrientationType = UiaEnum<
+        OrientationType,
+        winrt::Microsoft::UI::UIAutomation::AutomationOrientationType,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteOrientationType>;
+    using UiaOutlineStyles = UiaEnum<
+        OutlineStyles,
+        winrt::Microsoft::UI::UIAutomation::AutomationOutlineStyles,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteOutlineStyles>;
+    using UiaPatternId = UiaEnum<
+        PATTERNID,
+        winrt::Microsoft::UI::UIAutomation::AutomationPatternId,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemotePatternId>;
+    using UiaPropertyId = UiaEnum<
+        PROPERTYID,
+        winrt::Microsoft::UI::UIAutomation::AutomationPropertyId,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemotePropertyId>;
+    using UiaRowOrColumnMajor = UiaEnum<
+        RowOrColumnMajor,
+        winrt::Microsoft::UI::UIAutomation::AutomationRowOrColumnMajor,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteRowOrColumnMajor>;
+    using UiaSayAsInterpretAs = UiaEnum<
+        SayAsInterpretAs,
+        winrt::Microsoft::UI::UIAutomation::AutomationSayAsInterpretAs,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteSayAsInterpretAs>;
+    using UiaScrollAmount = UiaEnum<
+        ScrollAmount,
+        winrt::Microsoft::UI::UIAutomation::AutomationScrollAmount,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteScrollAmount>;
+    using UiaStyleId = UiaEnum<
+        int,
+        winrt::Microsoft::UI::UIAutomation::AutomationStyleId,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteStyleId>;
+    using UiaSupportedTextSelection = UiaEnum<
+        SupportedTextSelection,
+        winrt::Microsoft::UI::UIAutomation::AutomationSupportedTextSelection,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteSupportedTextSelection>;
+    using UiaSynchronizedInputType = UiaEnum<
+        SynchronizedInputType,
+        winrt::Microsoft::UI::UIAutomation::AutomationSynchronizedInputType,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteSynchronizedInputType>;
+    using UiaTextAttributeId = UiaEnum<
+        TEXTATTRIBUTEID,
+        winrt::Microsoft::UI::UIAutomation::AutomationTextAttributeId,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteTextAttributeId>;
+    using UiaTextDecorationLineStyle = UiaEnum<
+        TextDecorationLineStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationTextDecorationLineStyle,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteTextDecorationLineStyle>;
+    using UiaTextPatternRangeEndpoint = UiaEnum<
+        TextPatternRangeEndpoint,
+        winrt::Microsoft::UI::UIAutomation::AutomationTextPatternRangeEndpoint,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteTextPatternRangeEndpoint>;
+    using UiaTextUnit = UiaEnum<
+        TextUnit,
+        winrt::Microsoft::UI::UIAutomation::AutomationTextUnit,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteTextUnit>;
+    using UiaToggleState = UiaEnum<
+        ToggleState,
+        winrt::Microsoft::UI::UIAutomation::AutomationToggleState,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteToggleState>;
+    using UiaWindowInteractionState = UiaEnum<
+        WindowInteractionState,
+        winrt::Microsoft::UI::UIAutomation::AutomationWindowInteractionState,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteWindowInteractionState>;
+    using UiaWindowVisualState = UiaEnum<
+        WindowVisualState,
+        winrt::Microsoft::UI::UIAutomation::AutomationWindowVisualState,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteWindowVisualState>;
+    using UiaZoomUnit = UiaEnum<
+        ZoomUnit,
+        winrt::Microsoft::UI::UIAutomation::AutomationZoomUnit,
+        winrt::Microsoft::UI::UIAutomation::AutomationRemoteZoomUnit>;
